+++ src/main/java/de/samply/rstudio/group/RstudioGroupService.java
@@ -1,9 +1,0 @@
-package de.samply.rstudio.group;
-
-public interface RstudioGroupService {
-
-    void addUserToRstudioGroup(String email) throws RstudioGroupServiceException;
-
-    void removeUserFromRstudioGroup(String email) throws RstudioGroupServiceException;
-
-}